Removed compiler warning.

+++ apps/irc/irc.c
@@ -33,6 +33,9 @@
  * $Id$
  */
 
+#include <string.h>
+#include <stddef.h>
+
 #include "contiki-conf.h"
 #include "contiki.h"
 #include "contiki-net.h"
@@ -43,9 +46,6 @@
 #include "lib/ctk-textentry-cmdline.h"
 
 #include "lib/petsciiconv.h"
-
-
-#include <string.h>
 
 #ifdef IRC_CONF_WIDTH
 #define LOG_WIDTH IRC_CONF_WIDTH
@@ -274,7 +274,7 @@ PROCESS_THREAD(irc_process, ev, data)
       }
 #endif /* UIP_UDP */
     } else if(ev == ctk_signal_keypress) {
-      c = (ctk_arch_key_t)data;
+      c = (ctk_arch_key_t)(size_t)data;
       if(c == CH_ENTER) {
 	parse_line();
       } else {
